Use list index to identify fans and thermal zones
Previously, the label was used, but that would fail if a device had
multiple fans or thermal zones without a label. In that case, all of
those would refer to the first matching entry (and publish to the same
topic as well), so effectively only the first entry without label would
be available.

This changes the code to use the list index instead of the label to
identify the different sensors and autogenerating a label (e.g. fan1 or
temp1, matching lm-sensors output) to use for descriptive purposes and
for the MQTT topic (if no label is provided by the kernel).

This also fixes two small comment typos (thermal zone -> fan).

This fixes #73.